refactor: Extract Lunch Count admin configuration panel
Extract the inline configuration panel for Lunch Count from FeatureConfigurationPanel.tsx
into its own dedicated component, LunchCountConfigurationPanel.tsx, aligning with
the gold-standard BuildingConfigPanel pattern and updating the tracker accordingly.
